Update Three.js to fix a no matrices update bug
Our Three.js fork had a bug that invisible objects can lose
an opportunity to update their matrices and they can be rendered
at wrong place in the scene when it became visible. For example
due to this bug particles are rendered at wrong place indeed.

We have applied a patch to fix it on our Three.js fork end. With
this commit the updated Three.js will be imported to resolve the
problem.